requirements.txt
emoji==0.5.4
python-telegram-bot==21.10.0